+++ lib/gofer/base.rb
@@ -1,23 +1,27 @@
 module Gofer
   class Base
-    attr_reader :stdio
+    attr_reader :hostname, :username
+
+    KNOWN_OPTIONS = [
+      :stderr, :stdout, :stdio,
+      :output_prefix, :capture_exit_status, :quiet_stdout,
+      :quiet_stderr, :timeout
+    ]
 
     def initialize(opts = {})
-      @capture_exit_status = opts.delete(:capture_exit_status)
+      @capture_exit_status = opts[:capture_exit_status]
+      @stdio_class, @stdio_opts = opts[:stdio] || Stdio, {}
+      @timeout = opts[:timeout] || 12
 
-      quiet_stdout = opts.delete(:quiet_stdout)
-      output_prefix = opts.delete(:output_prefix)
-      quiet_stderr = opts.delete(:quiet_stderr)
-      stdout = opts[:stdout] || $stdout
-      stderr = opts[:stderr] || $stderr
+      @stdio_opts[:stdout] = opts[:stdout]
+      @stdio_opts[:quiet_stdout] = opts[:quiet_stdout]
+      @stdio_opts[:output_prefix] = opts[:output_prefix]
+      @stdio_opts[:quiet_stderr] = opts[:quiet_stderr]
+      @stdio_opts[:stderr] = opts[:stderr]
 
-      @stdio = (opts.delete(:stdio) || Stdio).new({
-        :stderr => stderr,
-        :quiet_stderr => quiet_stderr,
-        :output_prefix => output_prefix,
-        :quiet_stdout => quiet_stdout,
-        :stdout => stdout,
-      })
+      @ssh_opts = opts.delete_if do |k, v|
+        KNOWN_OPTIONS.include?(k)
+      end
     end
 
     # A wrapper that de-duplicates stuff to write to Stdio.
@@ -25,8 +29,8 @@ module Gofer
     def write_stdio(oe = {})
       oe[:output] << oe[:stderr_in] if oe[:stderr_in]
       oe[:output] << oe[:stdout_in] if oe[:stdout_in]
-      @stdio.stderr(oe[:stderr_in], oe[:opts]) if oe[:stderr_in]
-      @stdio.stdout(oe[:stdout_in], oe[:opts]) if oe[:stdout_in]
+      stdio.stderr(oe[:stderr_in], oe[:opts]) if oe[:stderr_in]
+      stdio.stdout(oe[:stdout_in], oe[:opts]) if oe[:stdout_in]
       oe[:stderr_out] << oe[:stderr_in] if oe[:stderr_in] && oe[:stderr_out]
       oe[:stdout_out] << oe[:stdout_in] if oe[:stdout_in] && oe[:stdout_out]
     end
@@ -34,11 +38,11 @@ module Gofer
     # Normalizes the opts so that we can accept opts from pretty much
     # anywhere (run, and others.)
 
-    def normalize_opts(opt = {})
-      opt[:timeout] = @timeout unless opt.has_key?(:timeout)
-      opt[:capture_exit_status] = @capture_exit_status unless \
-        opt.has_key?(:capture_exit_status)
-    opt
+    def normalize_opts(opts = {})
+      opts[:timeout] = @timeout unless opts.has_key?(:timeout)
+      opts[:capture_exit_status] = @capture_exit_status unless \
+        opts.has_key?(:capture_exit_status)
+      opts
     end
 
     # Allows you to see the +#to_s+ as "user@hostname" so that you can always
@@ -61,6 +65,11 @@ module Gofer
       if ! opts[:capture_exit_status] && out.exit_status != 0
         raise Error.new(self, out, command)
       end
+    end
+
+    private
+    def stdio
+      @stdio ||= @stdio_class.new(@stdio_opts)
     end
   end
 end