+++ spec/gofer/local_spec.rb
@@ -16,8 +16,8 @@ describe Gofer::Local do
   it "accepts a custom stdio" do
     host = Gofer::Local.new(:stdio => TempStdio)
     host.run("echo hello")
-    expect(host.stdio).to be_kind_of TempStdio
-    expect(host.stdio.stringio.string.strip).to eq "hello"
+    expect(host.send(:stdio)).to be_kind_of TempStdio
+    expect(host.send(:stdio).stringio.string.strip).to eq "hello"
   end
 
   describe :run do